登录

+++ src/main/java/org/ylc/frame/springboot/biz/entity/Department.java
@@ -24,7 +24,7 @@ public class Department implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty(value = "部门编号")
     @TableField("code")
@@ -36,20 +36,20 @@ public class Department implements Serializable {
 
     @ApiModelProperty(value = "上级部门ID")
     @TableField("pid")
-    private Integer pid;
+    private Long pid;
 
     @ApiModelProperty(value = "备注")
     @TableField("remark")
     private String remark;
 
     @TableField(value = "create_user", fill = FieldFill.INSERT)
-    private Integer createUser;
+    private Long createUser;
 
     @TableField(value = "create_time", fill = FieldFill.INSERT)
     private LocalDateTime createTime;
 
     @TableField(value = "update_user", fill = FieldFill.UPDATE)
-    private Integer updateUser;
+    private Long updateUser;
 
     @TableField(value = "update_time", fill = FieldFill.UPDATE)
     private LocalDateTime updateTime;
