+++ src/main/java/org/ylc/frame/springboot/biz/entity/Menu.java
@@ -24,11 +24,11 @@ public class Menu implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty(value = "上级ID")
     @TableField("pid")
-    private Integer pid;
+    private Long pid;
 
     @ApiModelProperty(value = "名称")
     @TableField("name")
@@ -71,13 +71,13 @@ public class Menu implements Serializable {
     private String remark;
 
     @TableField(value = "create_user", fill = FieldFill.INSERT)
-    private Integer createUser;
+    private Long createUser;
 
     @TableField(value = "create_time", fill = FieldFill.INSERT)
     private LocalDateTime createTime;
 
     @TableField(value = "update_user", fill = FieldFill.UPDATE)
-    private Integer updateUser;
+    private Long updateUser;
 
     @TableField(value = "update_time", fill = FieldFill.UPDATE)
     private LocalDateTime updateTime;
