+++ src/main/java/org/ylc/frame/springboot/biz/entity/RoleMenu.java
@@ -26,15 +26,15 @@ public class RoleMenu implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty(value = "角色ID")
     @TableField("role_id")
-    private Integer roleId;
+    private Long roleId;
 
     @ApiModelProperty(value = "菜单ID")
     @TableField("menu_id")
-    private Integer menuId;
+    private Long menuId;
 
 
 }
